more smooth implementation of slack message to calendar transfer + more (#43)
* better comments

* authentication for admin table

* updated links

* backend link change

* minor var changes

* change navbar

* change typo

* regex

* mortboot banner

* pom.xml and application.properties for email

* pom.xml again

* email stuff wip

* WORKING EMAIL SYSTEM

* removed mort boot banner

* banner

* approve method, comments, and encoded url

* using javauri for approvalLink

* more smooth implementation of slack message to calendar transfer

* minor var changes

* better comments

* authentication for admin table

* updated links

* backend link change

* regex

* change navbar

* change typo

* mortboot banner

* pom.xml and application.properties for email

* pom.xml again

* email stuff wip

* WORKING EMAIL SYSTEM

* removed mort boot banner

* banner

* approve method, comments, and encoded url

* more smooth implementation of slack message to calendar transfer

* using javauri for approvalLink

* origins

* personapicontroller update

* securityconfig update

---------

+++ src/main/java/com/nighthawk/spring_portfolio/mvc/bathroom/Issue.java
@@ -26,17 +26,17 @@ public class Issue {
     
     private int count;
 
-    private float positionX;
+    private double positionX;
 
-    private float positionY;
+    private double positionY;
 
-    public Issue(String bathroomName, String issue, int count, float positionX, float positionY)
-    {
-        this.bathroom = bathroomName;
-        this.issue = issue;
-        this.count = count;
-        this.positionX = positionX;
-        this.positionY = positionY;
+    public Issue(String bathroomName, String issue, int count, double positionX, double positionY)
+        {
+            this.bathroom = bathroomName;
+            this.issue = issue;
+            this.count = count;
+            this.positionX = positionX;
+            this.positionY = positionY;
 
     }
